+++ test/acceptance/tests/basic/terraform/public-listener-port-validate/main.tf
@@ -1,0 +1,18 @@
+provider "aws" {
+  region = "us-west-2"
+}
+
+variable "envoy_public_listener_port" {
+  type = number
+}
+
+module "test_client" {
+  source = "../../../../../../modules/mesh-task"
+  family = "family"
+  container_definitions = [{
+    name = "basic"
+  }]
+  retry_join                 = ["test"]
+  outbound_only              = true
+  envoy_public_listener_port = var.envoy_public_listener_port
+}